style: 💄 address PR comments
