+++ .github/workflows/main.yml
@@ -2,7 +2,7 @@ name: CI
 
 on:
   push:
-    branches: [master, beta, alpha]
+    branches: [master, beta, alpha, confidential-asset]
   pull_request:
     types: [assigned, opened, synchronize, reopened]
