+++ src/internal.ts
@@ -25,10 +25,7 @@ export { addConfidentialTransaction } from '~/api/procedures/addConfidentialTran
 export { addInstruction } from '~/api/procedures/addInstruction';
 export { executeManualInstruction } from '~/api/procedures/executeManualInstruction';
 export { executeConfidentialTransaction } from '~/api/procedures/executeConfidentialTransaction';
-export {
-  affirmConfidentialTransactions,
-  AffirmConfidentialTransactionParams,
-} from '~/api/procedures/affirmConfidentialTransactions';
+export { affirmConfidentialTransactions } from '~/api/procedures/affirmConfidentialTransactions';
 export { rejectConfidentialTransaction } from '~/api/procedures/rejectConfidentialTransaction';
 export {
   consumeAuthorizationRequests,
